+++ extensions/canvascast/src/utils/types.ts
@@ -12,7 +12,7 @@ export interface assignment {
   name: string;
   id: number;
   description: string;
-  date: any;
+  date: string;
   course: string;
   course_id: number;
   color: Color;
@@ -25,7 +25,7 @@ export interface announcement {
   course: string;
   id: number;
   markdown: string;
-  date: any;
+  date: string;
 }
 
 export interface modulesection {
@@ -45,6 +45,6 @@ export interface moduleitem {
 export interface Preferences {
   token: string;
   domain: string;
-  recent: boolean;
+  showRecent: boolean;
   numRecent: string;
 }