+++ extensions/easydict/src/types.ts
@@ -2,12 +2,13 @@
  * @author: tisfeng
  * @createTime: 2022-06-04 21:58
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-07-21 15:29
+ * @lastEditTime: 2022-07-31 22:48
  * @fileName: types.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
+import { Image } from "@raycast/api";
 import { TextTranslateResponse } from "tencentcloud-sdk-nodejs-tmt/tencentcloud/services/tmt/v20180321/tmt_models";
 import { LanguageDetectType } from "./detectLanguage";
 import { IcibaDictionaryResult } from "./dict/iciba/interface";
@@ -50,10 +51,10 @@ type RequestResult =
   | BaiduTranslateResult
   | TencentTranslateResult
   | CaiyunTranslateResult
-  | AppleTranslateResult
   | DeepLTranslateResult
   | IcibaDictionaryResult
-  | YoudaoDictionaryResult;
+  | YoudaoDictionaryResult
+  | string;
 
 export interface RequestErrorInfo {
   message: string;
@@ -120,13 +121,19 @@ export interface YoudaoTranslateReformatResultItem {
 export interface MyPreferences {
   language1: string;
   language2: string;
-  isAutomaticQuerySelectedText: boolean;
-  isAutomaticPlayWordAudio: boolean;
-  isDisplayTargetTranslationLanguage: boolean;
-  translationDisplayOrder: string;
+  enableAutomaticQuerySelectedText: boolean;
+  enableAutomaticPlayWordAudio: boolean;
+  enableDisplayTargetTranslationLanguage: boolean;
+  enableOpenInEudic: boolean;
+  translationSortOrder: string;
 
   youdaoAppId: string;
   youdaoAppSecret: string;
+
+  enableDeepLTranslate: boolean;
+  deepLAuthKey: string;
+
+  enableGoogleTranslate: boolean;
 
   enableBaiduTranslate: boolean;
   baiduAppId: string;
@@ -138,9 +145,6 @@ export interface MyPreferences {
 
   enableAppleLanguageDetect: boolean;
   enableAppleTranslate: boolean;
-
-  enableDeepLTranslate: boolean;
-  deepLAuthKey: string;
 
   enableCaiyunTranslate: boolean;
   caiyunToken: string;
@@ -189,10 +193,6 @@ export interface CaiyunTranslateResult {
   rc: string;
   target: string[];
   confidence: number;
-}
-
-export interface AppleTranslateResult {
-  translatedText: string;
 }
 
 /**
@@ -259,4 +259,11 @@ export interface QueryRecoredItem {
   timestamp: number;
   queryText: string;
   result?: string;
+}
+
+export interface WebTranslationItem {
+  type: QueryType;
+  webUrl: string;
+  icon: Image.ImageLike;
+  title: string;
 }