+++ extensions/emoji/src/emoji.tsx
@@ -1,4 +1,4 @@
-import { ActionPanel, List, closeMainWindow, popToRoot, getPreferenceValues, Action } from "@raycast/api";
+import { ActionPanel, List, getPreferenceValues, Action } from "@raycast/api";
 import { useState, useEffect } from "react";
 import type { ReactElement } from "react";
 import { createEmojiList } from "generate-emoji-list";
@@ -7,6 +7,13 @@ import { UnicodeVersion } from "generate-emoji-list/dist/createEmojiList";
 import emojiKeywords from "emojilib";
 import Fuse from "fuse.js";
 import { usePersistentState } from "raycast-toolkit";
+
+const { primaryAction, unicodeVersion, shortCodes } = getPreferenceValues<{
+  primaryAction: "paste" | "copy";
+  unicodeVersion: UnicodeVersion;
+  shortCodes: boolean;
+}>();
+const allEmojis = "All Emojis";
 
 type Category = { category: string; emojis: Emoji[] };
 type Emoji = {
@@ -27,6 +34,9 @@ const filterList = (list: Category[], searchText: string, category: string): Cat
   });
 };
 
+const getEmojipediaLink = (description: string) =>
+  `https://emojipedia.org/${description.toLowerCase().replace(/:? /g, "-")}/`;
+
 export default function Main(): ReactElement {
   const [list, setList] = usePersistentState<Emoji[]>("emoji-list-v2", []);
   const [categories, setCategories] = usePersistentState<string[]>("emoji-categories", []);
@@ -35,12 +45,10 @@ export default function Main(): ReactElement {
     // https://github.com/facebook/react/pull/22114
     let didUnmount = false;
 
-    const options = {
-      unicodeVersion: getPreferenceValues().unicodeVersion.value as UnicodeVersion,
-      features: { shortCodes: Boolean(getPreferenceValues().shortCodes.value) },
-    };
-
-    createEmojiList(options).then((list: Category[]) => {
+    createEmojiList({
+      unicodeVersion,
+      features: { shortCodes },
+    }).then((list: Category[]) => {
       if (!didUnmount) {
         setList(
           list.flatMap((category) =>
@@ -78,7 +86,7 @@ export default function Main(): ReactElement {
       onSearchTextChange={setSearchText}
       searchBarAccessory={
         <List.Dropdown tooltip="Select Category" onChange={setCategory}>
-          <List.Dropdown.Item key={category} title="All Emojis" value="" icon="🥳" />
+          <List.Dropdown.Item key={category} title={allEmojis} value="" icon="🥳" />
           {categories.map((category) => (
             <List.Dropdown.Item
               key={category}
@@ -94,46 +102,71 @@ export default function Main(): ReactElement {
         ? filterList(
             [
               !searchText && { category: "Recently Used", emojis: recentlyUsed },
-              { category: category || "Emojis", emojis: list },
+              { category: category || allEmojis, emojis: list },
             ].filter(Boolean) as Category[],
             searchText,
             category
           ).map((category: Category) => (
             <List.Section title={category.category} key={category.category}>
-              {category.emojis.map((emoji) => (
-                <List.Item
-                  key={emoji.description}
-                  id={`${category.category}${emoji.description}`}
-                  icon={emoji.emoji}
-                  title={emoji.description.replace(/\b(\w)/g, (s) => s.toUpperCase())}
-                  accessoryTitle={emoji?.shortCode?.join(" / ")}
-                  keywords={emoji.shortCode}
-                  actions={
-                    <ActionPanel>
-                      <ActionPanel.Section>
-                        <Action.Paste
-                          title="Paste Emoji in Active App"
-                          content={emoji.emoji}
-                          onPaste={() => {
-                            closeMainWindow();
-                            popToRoot();
-                            addToRecentlyUsed(emoji);
-                          }}
-                        />
-                        <Action.CopyToClipboard
-                          title="Copy Emoji to Clipboard"
-                          content={emoji.emoji}
-                          onCopy={() => {
-                            closeMainWindow();
-                            popToRoot();
-                            addToRecentlyUsed(emoji);
-                          }}
-                        />
-                      </ActionPanel.Section>
-                    </ActionPanel>
-                  }
-                />
-              ))}
+              {category.emojis.map((emoji) => {
+                const paste = (
+                  <Action.Paste
+                    content={emoji.emoji}
+                    onPaste={() => {
+                      addToRecentlyUsed(emoji);
+                    }}
+                  />
+                );
+                const copy = (
+                  <Action.CopyToClipboard
+                    content={emoji.emoji}
+                    onCopy={() => {
+                      addToRecentlyUsed(emoji);
+                    }}
+                  />
+                );
+                return (
+                  <List.Item
+                    key={emoji.description}
+                    id={`${category.category}${emoji.description}`}
+                    icon={emoji.emoji}
+                    title={emoji.description.replace(/\b(\w)/g, (s) => s.toUpperCase())}
+                    keywords={emoji.shortCode}
+                    actions={
+                      <ActionPanel>
+                        <ActionPanel.Section>
+                          {primaryAction === "paste" ? (
+                            <>
+                              {paste}
+                              {copy}
+                            </>
+                          ) : (
+                            <>
+                              {copy}
+                              {paste}
+                            </>
+                          )}
+                          {shortCodes && emoji.shortCode && (
+                            <Action.CopyToClipboard
+                              title="Copy Shortcode"
+                              content={emoji.shortCode[0]}
+                              onCopy={() => {
+                                addToRecentlyUsed(emoji);
+                              }}
+                            />
+                          )}
+                          <Action.OpenInBrowser title="View on Emojipedia" url={getEmojipediaLink(emoji.description)} />
+                        </ActionPanel.Section>
+                      </ActionPanel>
+                    }
+                    accessories={[
+                      {
+                        text: emoji?.shortCode?.join(" / "),
+                      },
+                    ]}
+                  />
+                );
+              })}
             </List.Section>
           ))
         : []}