+++ extensions/gopass/src/details.tsx
@@ -1,7 +1,7 @@
 import { Action, ActionPanel, Clipboard, closeMainWindow, Icon, List, showHUD, showToast, Toast } from "@raycast/api";
 import { useEffect, useState } from "react";
 import gopass from "./gopass";
-import { humanize } from "./utils";
+import { humanize, isValidUrl } from "./utils";
 import { copyPassword, pastePassword } from "./index";
 
 async function copy(key: string, value: string): Promise<void> {
@@ -15,13 +15,6 @@ async function paste(key: string, value: string): Promise<void> {
   await Clipboard.paste(value);
   await closeMainWindow();
 }
-
-const Actions = ({ copy, paste }: { copy: () => void; paste: () => void }): JSX.Element => (
-  <ActionPanel>
-    <Action title="Copy to Clipboard" icon={Icon.Clipboard} onAction={copy} />
-    <Action title="Paste to Active App" icon={Icon.Document} onAction={paste} />
-  </ActionPanel>
-);
 
 export default function ({ entry }: { entry: string }): JSX.Element {
   const [details, setDetails] = useState<string[]>([]);
@@ -40,27 +33,40 @@ export default function ({ entry }: { entry: string }): JSX.Element {
 
   return (
     <List isLoading={loading}>
-      {!loading && (
-        <List.Item
-          title="Password"
-          subtitle="*****************"
-          actions={<Actions copy={() => copyPassword(entry)} paste={() => pastePassword(entry)}></Actions>}
-        />
-      )}
-
-      {details.map((item, index) => {
-        const [key, ...values] = item.split(": ");
-        const value = values.join(": ");
-
-        return (
+      <List.Section title={"/" + entry}>
+        {!loading && (
           <List.Item
-            key={index}
-            title={humanize(key)}
-            subtitle={value}
-            actions={<Actions copy={() => copy(key, value)} paste={() => paste(key, value)}></Actions>}
+            title="Password"
+            subtitle="*****************"
+            actions={
+              <ActionPanel>
+                <Action title="Copy to Clipboard" icon={Icon.Clipboard} onAction={() => copyPassword(entry)} />
+                <Action title="Paste to Active App" icon={Icon.Document} onAction={() => pastePassword(entry)} />
+              </ActionPanel>
+            }
           />
-        );
-      })}
+        )}
+
+        {details.map((item, index) => {
+          const [key, ...values] = item.split(": ");
+          const value = values.join(": ");
+
+          return (
+            <List.Item
+              key={index}
+              title={humanize(key)}
+              subtitle={value}
+              actions={
+                <ActionPanel>
+                  {isValidUrl(value) && <Action.OpenInBrowser url={value} />}
+                  <Action title="Copy to Clipboard" icon={Icon.Clipboard} onAction={() => copy(key, value)} />
+                  <Action title="Paste to Active App" icon={Icon.Document} onAction={() => paste(key, value)} />
+                </ActionPanel>
+              }
+            />
+          );
+        })}
+      </List.Section>
     </List>
   );
 }