+++ extensions/gopass/src/index.tsx
@@ -2,6 +2,7 @@ import { Action, ActionPanel, Clipboard, closeMainWindow, Icon, List, showHUD, s
 import { useEffect, useState } from "react";
 import gopass from "./gopass";
 import Details from "./details";
+import { isDirectory } from "./utils";
 
 export async function copyPassword(entry: string): Promise<void> {
   try {
@@ -30,41 +31,61 @@ export async function pastePassword(entry: string): Promise<void> {
   }
 }
 
-export default function (): JSX.Element {
-  const [entries, setEntries] = useState<string[]>();
+const passwordActions = (entry: string) => (
+  <>
+    <Action title="Copy Password to Clipboard" icon={Icon.Clipboard} onAction={() => copyPassword(entry)} />
+    <Action
+      title="Paste Password to Active App"
+      icon={Icon.Document}
+      onAction={() => pastePassword(entry)}
+      shortcut={{ modifiers: ["cmd", "shift"], key: "enter" }}
+    />
+  </>
+);
+
+const getIcon = (entry: string) => (isDirectory(entry) ? Icon.Folder : Icon.Key);
+
+const getTarget = (entry: string) => (isDirectory(entry) ? <Main prefix={entry} /> : <Details entry={entry} />);
+
+export default function Main({ prefix = "" }): JSX.Element {
+  const [entries, setEntries] = useState<string[]>([]);
   const [loading, setLoading] = useState<boolean>(true);
+  const [searchText, setSearchText] = useState("");
 
   useEffect((): void => {
     gopass
-      .list()
+      .list({ limit: searchText ? -1 : 0, prefix, directoriesFirst: true, stripPrefix: true })
+      .then((data) => data.filter((item) => item.toLowerCase().includes(searchText.toLowerCase())))
       .then(setEntries)
       .catch(async (error) => {
         console.error(error);
         await showToast({ title: "Could not load passwords", style: Toast.Style.Failure });
       })
       .finally(() => setLoading(false));
-  }, []);
+  }, [searchText]);
 
   return (
-    <List isLoading={loading}>
-      {entries?.map((entry, i) => (
-        <List.Item
-          key={i}
-          title={entry}
-          actions={
-            <ActionPanel>
-              <Action.Push title="Show Details" icon={Icon.Document} target={<Details entry={entry} />} />
-              <Action title="Copy Password to Clipboard" icon={Icon.Clipboard} onAction={() => copyPassword(entry)} />
-              <Action
-                title="Paste Password to Active App"
-                icon={Icon.Document}
-                onAction={() => pastePassword(entry)}
-                shortcut={{ modifiers: ["cmd", "shift"], key: "enter" }}
-              />
-            </ActionPanel>
-          }
-        />
-      ))}
+    <List isLoading={loading} enableFiltering={false} onSearchTextChange={setSearchText}>
+      <List.Section title={searchText ? "Results" : "/" + prefix} subtitle={searchText && String(entries.length)}>
+        {entries.map((entry, i) => {
+          const fullPath = prefix + entry;
+
+          return (
+            <List.Item
+              key={i}
+              title={entry}
+              icon={getIcon(entry)}
+              accessories={[{ icon: Icon.ChevronRight }]}
+              actions={
+                <ActionPanel>
+                  <Action.Push title="Show Details" icon={getIcon(entry)} target={getTarget(fullPath)} />
+                  {!isDirectory(entry) && passwordActions(fullPath)}
+                </ActionPanel>
+              }
+            />
+          );
+        })}
+      </List.Section>
     </List>
   );
 }