+++ extensions/linear/src/helpers/withLinearClient.tsx
@@ -1,19 +1,18 @@
-import { useEffect, useState } from "react";
+import { useMemo, useState } from "react";
 import { Detail } from "@raycast/api";
 import { LinearClient, LinearGraphQLClient } from "@linear/sdk";
 
-import { authorize, oauthClient } from "../api/oauth";
+import { authorize } from "../api/oauth";
 
 let linearClient: LinearClient | null = null;
 
 export function withLinearClient(component: JSX.Element) {
   const [x, forceRerender] = useState(0);
 
-  useEffect(() => {
+  // we use a `useMemo` instead of `useEffect` to avoid a render
+  useMemo(() => {
     (async function () {
-      const tokens = await oauthClient.getTokens();
-
-      const accessToken = tokens?.accessToken || (await authorize());
+      const accessToken = await authorize();
 
       linearClient = new LinearClient({ accessToken });