+++ extensions/wakatime/src/hooks/index.ts
@@ -1,6 +1,7 @@
-export { useActivityChange } from "./use-activity-change";
-export { useLeaderBoard } from "./use-leaderboard";
-export { usePrivateLeaderBoards } from "./use-private-leaderboards";
-export { useProjects } from "./use-projects";
-export { useSummary } from "./use-summary";
 export { useUser } from "./use-user";
+export { useSummary } from "./use-summary";
+export { useProjects } from "./use-projects";
+export { useLeaderBoard } from "./use-leaderboard";
+export { useTodaySummary } from "./use-today-summary";
+export { useActivityChange } from "./use-activity-change";
+export { usePrivateLeaderBoards } from "./use-private-leaderboards";