+++ extensions/xcode/src/services/xcode-cleanup.service.ts
@@ -7,21 +7,21 @@ export class XcodeCleanupService {
   /**
    * Clear the Xcode Derived Data directory
    */
-  clearDerivedData(): Promise<void> {
+  static clearDerivedData(): Promise<void> {
     return execAsync("rm -rf ~/Library/Developer/Xcode/DerivedData").then();
   }
 
   /**
    * Clear Swift Package Manager Cache
    */
-  clearSwiftPackageManagerCache(): Promise<void> {
+  static clearSwiftPackageManagerCache(): Promise<void> {
     return execAsync("rm -rf ~/Library/Caches/org.swift.swiftpm/repositories").then();
   }
 
   /**
    * Delete unsupported Simulators
    */
-  deleteUnsupportedSimulators(): Promise<void> {
+  static deleteUnsupportedSimulators(): Promise<void> {
     return execAsync("xcrun simctl delete unavailable").then();
   }
 }