+++ extensions/unsplash/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unsplash Changelog
 
+## [Update] - 2022-08-04
+
+- Updated Raycast API to 1.38.2
+- Added option to update random images.
+
 ## [Update] - 2022-06-28
 
 - Updated Raycast API to 1.36.1