+++ extensions/wakatime/CHANGELOG.md
@@ -14,4 +14,8 @@
 - Fixed Sorting of Recent Projects.
 - Fixed `NaN%` in activity change
 
+## [Add Today Summary Menu Bar Command] - 2022-07-30
+
+- 🚀 Add Today Summary Menu Bar Command
+
 ## [Initial Version] - 2022-05-04